+++ HISTORY.md
@@ -9,6 +9,8 @@ Twig
   (GH-7. Thanks [joelmoss](https://github.com/joelmoss)!)
 * FIX: Suppress `which` errors.
   (GH-9. Thanks [badboy](https://github.com/badboy)!)
+* FIX: Exit with a non-zero status when trying to get a branch property that
+  doesn't exist.
 * FIX: In list view, render line breaks (in properties) as spaces.
 
 1.0 (2013-02-05)